Update README.md
This project focuses on analyzing the results of the 2024 General Elections in India. It uses SQL queries to manage and analyze data related to constituencies, candidates, political parties, and election outcomes. The goal is to perform detailed analysis and generate insights, such as the total number of seats won by alliances, distribution of votes, and comparison of results across states.